Rework the Docker packaging
The Docker packaging this was taken from did not build, primarily due to
various problems with the django settings.py file. The Django settings
were augmented just enough to run successfully by adding a valid
SECRET_KEY setting. As shown by actually running the generated image,
the database setup needs to happen before the webapp will actually run
successfully. Just enough was done to get it to run and produce a 500
error.

The image was ported to CentOS 7, where several useful packages are
available, as compared to CentOS 6. The Dockerfile was rearranged so
that the least amount of work is needed during rebuild for typical
cases.

The Makefile was simplified so that various envsubst commands did not
overwrite parts of the Dockerfile that should have stayed present.
Various confusion between make syntax and shell syntax were fixed.

The decision was made to put both the nginx and uwsgi processes into the
same Docker container for ease of management and to reduce filesystem
duplication. The supervisord daemon is used to manage the processes.

The nginx and uwsgi configurations were changed to log to stdout/stderr
natively, rather than going through several "tail" processes. The
"docker logs" command continues to work, and several fewer processes are
necessary.

The directory structure was rearranged so that all of the webapp code
lives in one directory, and all of the other configuration lives in
another directory. This helps to keep things organized.

Notes were added about NFS mounts.